\name{Attributes_40}
\alias{Attributes_40}
\docType{data}
\title{
Attributes for the Showers' card-sorting task (40 items)
}
\description{
This is a set of adjectives available to participants completing the Showers' card-sorting task. The task was developed by Linville (1985, 1987) and then adapted by Showers (1992). This set is commonly used in the literature, and it contains 20 positive and 20 negative adjectives.
}
\usage{data("Attributes_40")}
\format{
  A vector of 40 strings.
}
\source{
Showers, C. (1992). Compartmentalization of positive and negative
self-knowledge: Keeping bad apples out of the bunch. Journal of
Personality and Social Psychology, 62, 1036-1049. https://doi.org/10.1037/0022-3514.62.6.1036
}
\references{
Linville, E W. (1985). Self-complexity and affective extremity: Don't
put all of your eggs in one cognitive basket. Social Cognition, 3, 94-
120. https://doi.org/10.1521/soco.1985.3.1.94
Linville, E W. (1987). Self-complexity as a cognitive buffer against
stress-related illness and depression. Journal of Personality and Social Psychology, 52, 663-676.https://doi.org/10.1037/0022-3514.52.4.663
}
\examples{
data(Attributes_40)
str(Attributes_40)
}
\keyword{datasets}
