% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_H.R
\name{calculate_H}
\alias{calculate_H}
\title{Calculate the H index of self-complexity (dimensionality measure)}
\usage{
calculate_H(data, att_column, id_column, vector)
}
\arguments{
\item{x}{factor}
}
\value{
A tibble
}
\description{
Calculate the H index of self-complexity (dimensionality measure)
}
\examples{
fcount(iris$Species)
}
