% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_phi.R
\name{calculate_phi}
\alias{calculate_phi}
\title{Calculate phi index}
\usage{
calculate_phi(data, att_column, id_column, pos_att_vector, neg_att_vector)
}
\arguments{
\item{data}{A data.frame like object, see details for more information}

\item{att_column}{unquoted variable name of a single column. See details for
structure of this variable.}

\item{id_column}{unquoted variable name of a single column. The participant's unique
identifier.}

\item{pos_att_vector}{A vector of positive attributes available to the participant.}

\item{neg_att_vector}{A vector of negative attributes available to the participant.}
}
\value{
A tibble
}
\description{
Calculate phi index
}
\examples{
library(selfcomplexity)
data(Showers_onesort, package = "selfcomplexity")
data(Attributes_df, package = "selfcomplexity")
pos <- Attributes_df \%>\%
dplyr::filter(Positive == 1) \%>\%
dplyr::pull(Attribute)
neg <- Attributes_df \%>\%
dplyr::filter(Negative == 1) \%>\%
dplyr::pull(Attribute)
}
