% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_checks.r
\name{check_columns_exist}
\alias{check_columns_exist}
\title{Check columns exist}
\usage{
check_columns_exist(data, ...)
}
\arguments{
\item{data}{a data.frame like object}

\item{...}{unquoted variable names}
}
\description{
Ensure that specified columns are present in the data.frame
like object
}
