% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_checks.r
\name{check_for_low_values}
\alias{check_for_low_values}
\title{Check for low values}
\usage{
check_for_low_values(my_list)
}
\arguments{
\item{my_list}{a list or vector}
}
\description{
Check that the list-like object does not contain
numbers 0-2 (as per Phi calculation conditions).
Returns a 1 if low values present, and 0 otherwise.
}
