% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_checks.R
\name{check_input_data}
\alias{check_input_data}
\title{Check input data}
\usage{
check_input_data(data)
}
\arguments{
\item{data}{a data.frame like object}
}
\description{
A utility function to ensure that the input data.frame like
object for the various functions is appropriate
}
