% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_unique_attributes}
\alias{extract_unique_attributes}
\title{Extract unique attributes}
\usage{
extract_unique_attributes(att)
}
\arguments{
\item{att}{character vector of comma separated words
(typically attributes in this context)}
}
\value{
character vector of unique words
}
\description{
From a vector of characters in which each element
is a comma separated collection of words find all
the unique words.
}
