% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/into_symbol.R
\name{into_symbol}
\alias{into_symbol}
\title{Turn string/symbol into a symbol}
\usage{
into_symbol(value)
}
\arguments{
\item{value}{a single string or symbol}
}
\description{
A utility function to ensure that whether user submits column names as
strings or symbols, they work appropriately in the key functions
}
