\name{minimal_example}
\alias{minimal_example}
\docType{data}
\title{
Minimal example of participant with 3 self-aspects
}
\description{
This is a subset of the complexity_data dataset, and contains just one participant, with three self-aspects. Given the simplicity of this case, indices can be calculated by hand and used as test cases.
}
\usage{data("minimal_example")}
\format{
  A data frame with 3 observations on the following 3 variables.
  \describe{
    \item{\code{ResponseId}}{a character vector}
    \item{\code{Subtype_name}}{a character vector}
    \item{\code{Attributes}}{a character vector}
  }
}

\examples{
data(minimal_example)

}
\keyword{datasets}
