% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_H.R
\name{calculate_H}
\alias{calculate_H}
\title{Calculate the H index of self-complexity (dimensionality measure)}
\usage{
calculate_H(data, att_column, id_column, vector)
}
\arguments{
\item{x}{factor}
}
\value{
A tibble
}
\description{
Calculate the H index of self-complexity (dimensionality measure)
}
\examples{

All_Attributes <- c("Capable", "Comfortable", "Communicative", "Confident", "Disagreeing",
"Disorganised",
"Energetic", "Friendly", "Fun and Entertaining", "Giving", "Happy", "Hardworking",
"Hopeless", "Immature", "Incompetent", "Indecisive", "Independent", "Inferior",
"Insecure", "Intelligent", "Interested", "Irresponsible", "Irritable", "Isolated",
"Lazy", "Like a failure", "Lovable", "Mature", "Needed", "Optimistic", "Organised",
"Outgoing", "Sad and Blue", "Self-centered", "Successful", "Tense", "Uncomfortable",
"Unloved", "Weary", "Worthless")

calculate_H(complexity_data, att_column = "Attributes", id_column = "ResponseId",
vector = All_Attributes)
}
