% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_h.R
\name{calculate_H}
\alias{calculate_H}
\title{Calculate the H index of self-complexity (dimensionality measure)}
\usage{
calculate_H(data, att_column, id_column, vector)
}
\arguments{
\item{data}{A data.frame like object, see details for more information}

\item{att_column}{unquoted variable name of a single column. See details for
structure of this variable.}

\item{id_column}{unquoted variable name of a single column. The partipants unique
identifier}

\item{vector}{character vector of all attributes.}
}
\value{
A tibble
}
\description{
Longer description
}
\details{
Further details on some of the arguments is given below.
\itemize{
\item{"data"}{
The expectation is that we have tidy, long format data. Each
participant has a unique identifier, described by the id_column
argument. There may be several observations for single participants
in which case they represent different subtypes. It may be convenient
to keep the subtype identification column in the data but it is not required
for the purpose of this calculation.
}
\item{"att_column"}{
At present we expect a column where each observation is a single,
comma separated, character value of multiple adjectives, e.g
"Comfortable,Disorganised,Interested,Irresponsible"
}
}
}
\examples{
library(selfcomplexity)
data(complexity_data, package = "selfcomplexity")
data(attributes_40, package = "selfcomplexity")
calculate_H(complexity_data, Attributes, ResponseId, attributes_40)
}
