% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_overlap.R
\name{calculate_overlap}
\alias{calculate_overlap}
\title{Calculate overlap index}
\usage{
calculate_overlap(data, na_name_rm = TRUE)
}
\arguments{
\item{data}{A data.frame like object, see details for more information}

\item{na_name_rm}{Boolean variable to determine whether empty aspect names
should be removed}
}
\value{
A tibble
}
\description{
Calculate overlap index
}
\examples{
library(selfcomplexity)
data(complexity_data, package = "selfcomplexity")
calculate_overlap(complexity_data, na_name_rm = TRUE)
}
