% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_overlap.R
\name{calculate_overlap}
\alias{calculate_overlap}
\title{Calculate overlap index}
\usage{
calculate_overlap(subject_df, na.name.rm = TRUE)
}
\arguments{
\item{na.name.rm}{}
}
\value{
dataframe
}
\description{
Calculate overlap index
}
\examples{
calculate_overlap(example_data, na.name.rm = TRUE)
}
