% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexity_data.R
\docType{data}
\name{complexity_data}
\alias{complexity_data}
\title{Data from the Showers card sort task}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\source{
\href{https://phenome.jax.org/projects/Moore1b}{QTL Archive}
}
\usage{
data(selfcomplexity)
}
\description{
Data from a study that used the Showers card sort task
and then calculated self-complexity indices
}
\examples{
data(complexity_data)
calculate_H(complexity_data)
}
\references{
Banas and Smyth (2021). Structure, Content and Inter‐relationships
between Self‐aspects: Integrating Findings from the Social Identity and
Self Complexity Traditions. European Journal of Social Psychology.
(\href{https://doi.org/10.1002/ejsp.2760}{Online article})
}
\keyword{datasets}
